extract command handlers into separate command module
