+++ crates/goose-cli/src/session/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod context;
+pub mod extension;
+pub mod plan;
+pub mod prompt;
+pub mod recipe;
+pub mod settings;
