+++ crates/goose-cli/src/session/mod.rs
@@ -1,4 +1,5 @@
 mod builder;
+mod commands;
 mod completion;
 mod export;
 mod input;
@@ -14,24 +15,20 @@ use goose::agents::AgentEvent;
 use goose::permission::permission_confirmation::PrincipalType;
 use goose::permission::Permission;
 use goose::permission::PermissionConfirmation;
-use goose::providers::base::Provider;
 pub use goose::session::Identifier;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use completion::GooseCompleter;
 use etcetera::{choose_app_strategy, AppStrategy};
-use goose::agents::extension::{Envs, ExtensionConfig};
 use goose::agents::{Agent, SessionConfig};
 use goose::config::Config;
 use goose::message::{Message, MessageContent};
 use goose::session;
 use input::InputResult;
 use mcp_core::handler::ToolError;
-use mcp_core::prompt::PromptMessage;
 use mcp_core::protocol::JsonRpcMessage;
 use mcp_core::protocol::JsonRpcNotification;
 
-use rand::{distributions::Alphanumeric, Rng};
 use serde_json::Value;
 use std::collections::HashMap;
 use std::path::PathBuf;
@@ -106,180 +103,6 @@ impl Session {
             scheduled_job_id,
             max_turns,
         }
-    }
-
-    /// Helper function to summarize context messages
-    async fn summarize_context_messages(
-        messages: &mut Vec<Message>,
-        agent: &Agent,
-        message_suffix: &str,
-    ) -> Result<()> {
-        // Summarize messages to fit within context length
-        let (summarized_messages, _) = agent.summarize_context(messages).await?;
-        let msg = format!("Context maxed out\n{}\n{}", "-".repeat(50), message_suffix);
-        output::render_text(&msg, Some(Color::Yellow), true);
-        *messages = summarized_messages;
-
-        Ok(())
-    }
-
-    /// Add a stdio extension to the session
-    ///
-    /// # Arguments
-    /// * `extension_command` - Full command string including environment variables
-    ///   Format: "ENV1=val1 ENV2=val2 command args..."
-    pub async fn add_extension(&mut self, extension_command: String) -> Result<()> {
-        let mut parts: Vec<&str> = extension_command.split_whitespace().collect();
-        let mut envs = std::collections::HashMap::new();
-
-        // Parse environment variables (format: KEY=value)
-        while let Some(part) = parts.first() {
-            if !part.contains('=') {
-                break;
-            }
-            let env_part = parts.remove(0);
-            let (key, value) = env_part.split_once('=').unwrap();
-            envs.insert(key.to_string(), value.to_string());
-        }
-
-        if parts.is_empty() {
-            return Err(anyhow::anyhow!("No command provided in extension string"));
-        }
-
-        let cmd = parts.remove(0).to_string();
-        // Generate a random name for the ephemeral extension
-        let name: String = rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(8)
-            .map(char::from)
-            .collect();
-
-        let config = ExtensionConfig::Stdio {
-            name,
-            cmd,
-            args: parts.iter().map(|s| s.to_string()).collect(),
-            envs: Envs::new(envs),
-            env_keys: Vec::new(),
-            description: Some(goose::config::DEFAULT_EXTENSION_DESCRIPTION.to_string()),
-            // TODO: should set timeout
-            timeout: Some(goose::config::DEFAULT_EXTENSION_TIMEOUT),
-            bundled: None,
-        };
-
-        self.agent
-            .add_extension(config)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to start extension: {}", e))?;
-
-        // Invalidate the completion cache when a new extension is added
-        self.invalidate_completion_cache().await;
-
-        Ok(())
-    }
-
-    /// Add a remote extension to the session
-    ///
-    /// # Arguments
-    /// * `extension_url` - URL of the server
-    pub async fn add_remote_extension(&mut self, extension_url: String) -> Result<()> {
-        let name: String = rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(8)
-            .map(char::from)
-            .collect();
-
-        let config = ExtensionConfig::Sse {
-            name,
-            uri: extension_url,
-            envs: Envs::new(HashMap::new()),
-            env_keys: Vec::new(),
-            description: Some(goose::config::DEFAULT_EXTENSION_DESCRIPTION.to_string()),
-            // TODO: should set timeout
-            timeout: Some(goose::config::DEFAULT_EXTENSION_TIMEOUT),
-            bundled: None,
-        };
-
-        self.agent
-            .add_extension(config)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to start extension: {}", e))?;
-
-        // Invalidate the completion cache when a new extension is added
-        self.invalidate_completion_cache().await;
-
-        Ok(())
-    }
-
-    /// Add a builtin extension to the session
-    ///
-    /// # Arguments
-    /// * `builtin_name` - Name of the builtin extension(s), comma separated
-    pub async fn add_builtin(&mut self, builtin_name: String) -> Result<()> {
-        for name in builtin_name.split(',') {
-            let config = ExtensionConfig::Builtin {
-                name: name.trim().to_string(),
-                display_name: None,
-                // TODO: should set a timeout
-                timeout: Some(goose::config::DEFAULT_EXTENSION_TIMEOUT),
-                bundled: None,
-            };
-            self.agent
-                .add_extension(config)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to start builtin extension: {}", e))?;
-        }
-
-        // Invalidate the completion cache when a new extension is added
-        self.invalidate_completion_cache().await;
-
-        Ok(())
-    }
-
-    pub async fn list_prompts(
-        &mut self,
-        extension: Option<String>,
-    ) -> Result<HashMap<String, Vec<String>>> {
-        let prompts = self.agent.list_extension_prompts().await;
-
-        // Early validation if filtering by extension
-        if let Some(filter) = &extension {
-            if !prompts.contains_key(filter) {
-                return Err(anyhow::anyhow!("Extension '{}' not found", filter));
-            }
-        }
-
-        // Convert prompts into filtered map of extension names to prompt names
-        Ok(prompts
-            .into_iter()
-            .filter(|(ext, _)| extension.as_ref().is_none_or(|f| f == ext))
-            .map(|(extension, prompt_list)| {
-                let names = prompt_list.into_iter().map(|p| p.name).collect();
-                (extension, names)
-            })
-            .collect())
-    }
-
-    pub async fn get_prompt_info(&mut self, name: &str) -> Result<Option<output::PromptInfo>> {
-        let prompts = self.agent.list_extension_prompts().await;
-
-        // Find which extension has this prompt
-        for (extension, prompt_list) in prompts {
-            if let Some(prompt) = prompt_list.iter().find(|p| p.name == name) {
-                return Ok(Some(output::PromptInfo {
-                    name: prompt.name.clone(),
-                    description: prompt.description.clone(),
-                    arguments: prompt.arguments.clone(),
-                    extension: Some(extension),
-                }));
-            }
-        }
-
-        Ok(None)
-    }
-
-    pub async fn get_prompt(&mut self, name: &str, arguments: Value) -> Result<Vec<PromptMessage>> {
-        let result = self.agent.get_prompt(name, arguments).await?;
-        Ok(result.messages)
     }
 
     /// Process a single message and get the response
@@ -383,188 +206,6 @@ impl Session {
                 eprintln!("Warning: Failed to load command history: {}", err);
             }
         }
-    }
-
-    /// Handle normal mode message processing in interactive session
-    async fn handle_normal_mode_message(&mut self, content: &str) -> Result<()> {
-        self.add_user_message_to_history(content).await?;
-        self.update_project_tracker(content)?;
-
-        output::show_thinking();
-        self.process_agent_response(true).await?;
-        output::hide_thinking();
-
-        Ok(())
-    }
-
-    /// Handle plan mode message processing in interactive session
-    async fn handle_plan_mode_message(&mut self, content: &str) -> Result<()> {
-        let mut plan_messages = self.messages.clone();
-        plan_messages.push(Message::user().with_text(content));
-        let reasoner = get_reasoner()?;
-        self.plan_with_reasoner_model(plan_messages, reasoner)
-            .await?;
-
-        Ok(())
-    }
-
-    /// Handle theme toggle command
-    fn handle_theme_toggle(&self) {
-        let current = output::get_theme();
-        let new_theme = match current {
-            output::Theme::Light => {
-                println!("Switching to Dark theme");
-                output::Theme::Dark
-            }
-            output::Theme::Dark => {
-                println!("Switching to Ansi theme");
-                output::Theme::Ansi
-            }
-            output::Theme::Ansi => {
-                println!("Switching to Light theme");
-                output::Theme::Light
-            }
-        };
-        output::set_theme(new_theme);
-    }
-
-    /// Handle goose mode setting command
-    fn handle_goose_mode_setting(&self, mode: &str) -> Result<bool> {
-        let config = Config::global();
-        let mode = mode.to_lowercase();
-
-        // Check if mode is valid
-        if !["auto", "approve", "chat", "smart_approve"].contains(&mode.as_str()) {
-            output::render_error(&format!(
-                "Invalid mode '{}'. Mode must be one of: auto, approve, chat",
-                mode
-            ));
-            return Ok(false);
-        }
-
-        config
-            .set_param("GOOSE_MODE", Value::String(mode.to_string()))
-            .unwrap();
-        output::goose_mode_message(&format!("Goose mode set to '{}'", mode));
-
-        Ok(true)
-    }
-
-    /// Handle plan command
-    async fn handle_plan_command(&mut self, options: input::PlanCommandOptions) -> Result<()> {
-        self.run_mode = RunMode::Plan;
-        output::render_enter_plan_mode();
-
-        let message_text = options.message_text;
-        if !message_text.is_empty() {
-            let mut plan_messages = self.messages.clone();
-            plan_messages.push(Message::user().with_text(&message_text));
-
-            let reasoner = get_reasoner()?;
-            self.plan_with_reasoner_model(plan_messages, reasoner)
-                .await?;
-        }
-
-        Ok(())
-    }
-
-    /// Handle clear command
-    fn handle_clear_command(&mut self) {
-        self.messages.clear();
-        tracing::info!("Chat context cleared by user.");
-        output::render_message(
-            &Message::assistant().with_text("Chat context cleared."),
-            self.debug,
-        );
-    }
-
-    /// Handle recipe generation command
-    async fn handle_recipe_command(&mut self, filepath_opt: Option<String>) -> Result<()> {
-        println!("{}", console::style("Generating Recipe").green());
-
-        output::show_thinking();
-        let recipe = self.agent.create_recipe(self.messages.clone()).await;
-        output::hide_thinking();
-
-        match recipe {
-            Ok(recipe) => {
-                // Use provided filepath or default
-                let filepath_str = filepath_opt.as_deref().unwrap_or("recipe.yaml");
-                match self.save_recipe(&recipe, filepath_str) {
-                    Ok(path) => println!(
-                        "{}",
-                        console::style(format!("Saved recipe to {}", path.display())).green()
-                    ),
-                    Err(e) => {
-                        println!("{}", console::style(e).red());
-                    }
-                }
-            }
-            Err(e) => {
-                println!(
-                    "{}: {:?}",
-                    console::style("Failed to generate recipe").red(),
-                    e
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle summarize command
-    async fn handle_summarize_command(&mut self) -> Result<()> {
-        let prompt = "Are you sure you want to summarize this conversation? This will condense the message history.";
-        let should_summarize = match cliclack::confirm(prompt).initial_value(true).interact() {
-            Ok(choice) => choice,
-            Err(e) => {
-                if e.kind() == std::io::ErrorKind::Interrupted {
-                    false // If interrupted, set should_summarize to false
-                } else {
-                    return Err(e.into());
-                }
-            }
-        };
-
-        if should_summarize {
-            println!("{}", console::style("Summarizing conversation...").yellow());
-            output::show_thinking();
-
-            // Get the provider for summarization
-            let provider = self.agent.provider().await?;
-
-            // Call the summarize_context method which uses the summarize_messages function
-            let (summarized_messages, _) = self.agent.summarize_context(&self.messages).await?;
-
-            // Update the session messages with the summarized ones
-            self.messages = summarized_messages;
-
-            // Persist the summarized messages
-            if let Some(session_file) = &self.session_file {
-                session::persist_messages_with_schedule_id(
-                    session_file,
-                    &self.messages,
-                    Some(provider),
-                    self.scheduled_job_id.clone(),
-                )
-                .await?;
-            }
-
-            output::hide_thinking();
-            println!(
-                "{}",
-                console::style("Conversation has been summarized.").green()
-            );
-            println!(
-                "{}",
-                console::style("Key information has been preserved while reducing context length.")
-                    .green()
-            );
-        } else {
-            println!("{}", console::style("Summarization cancelled.").yellow());
-        }
-
-        Ok(())
     }
 
     /// Start an interactive session, optionally with an initial message
@@ -683,83 +324,6 @@ impl Session {
                 .map(|p| format!(" Recorded to {}", p.display()))
                 .unwrap_or_default()
         );
-        Ok(())
-    }
-
-    async fn plan_with_reasoner_model(
-        &mut self,
-        plan_messages: Vec<Message>,
-        reasoner: Arc<dyn Provider>,
-    ) -> Result<(), anyhow::Error> {
-        let plan_prompt = self.agent.get_plan_prompt().await?;
-        output::show_thinking();
-        let (plan_response, _usage) = reasoner.complete(&plan_prompt, &plan_messages, &[]).await?;
-        output::render_message(&plan_response, self.debug);
-        output::hide_thinking();
-        let planner_response_type =
-            classify_planner_response(plan_response.as_concat_text(), self.agent.provider().await?)
-                .await?;
-
-        match planner_response_type {
-            PlannerResponseType::Plan => {
-                println!();
-                let should_act = match cliclack::confirm(
-                    "Do you want to clear message history & act on this plan?",
-                )
-                .initial_value(true)
-                .interact()
-                {
-                    Ok(choice) => choice,
-                    Err(e) => {
-                        if e.kind() == std::io::ErrorKind::Interrupted {
-                            false // If interrupted, set should_act to false
-                        } else {
-                            return Err(e.into());
-                        }
-                    }
-                };
-                if should_act {
-                    output::render_act_on_plan();
-                    self.run_mode = RunMode::Normal;
-                    // set goose mode: auto if that isn't already the case
-                    let config = Config::global();
-                    let curr_goose_mode =
-                        config.get_param("GOOSE_MODE").unwrap_or("auto".to_string());
-                    if curr_goose_mode != "auto" {
-                        config
-                            .set_param("GOOSE_MODE", Value::String("auto".to_string()))
-                            .unwrap();
-                    }
-
-                    // clear the messages before acting on the plan
-                    self.messages.clear();
-                    // add the plan response as a user message
-                    let plan_message = Message::user().with_text(plan_response.as_concat_text());
-                    self.messages.push(plan_message);
-                    // act on the plan
-                    output::show_thinking();
-                    self.process_agent_response(true).await?;
-                    output::hide_thinking();
-
-                    // Reset run & goose mode
-                    if curr_goose_mode != "auto" {
-                        config
-                            .set_param("GOOSE_MODE", Value::String(curr_goose_mode.to_string()))
-                            .unwrap();
-                    }
-                } else {
-                    // add the plan response (assistant message) & carry the conversation forward
-                    // in the next round, the user might wanna slightly modify the plan
-                    self.messages.push(plan_response);
-                }
-            }
-            PlannerResponseType::ClarifyingQuestions => {
-                // add the plan response (assistant message) & carry the conversation forward
-                // in the next round, the user will answer the clarifying questions
-                self.messages.push(plan_response);
-            }
-        }
-
         Ok(())
     }
 
@@ -907,7 +471,7 @@ impl Session {
                                         } else {
                                             "Goose automatically summarized messages to continue processing."
                                         };
-                                        Self::summarize_context_messages(&mut self.messages, &self.agent, message_suffix).await?;
+                                        crate::session::commands::context::summarize_context_messages(&mut self.messages, &self.agent, message_suffix).await?;
                                     }
                                     _ => {
                                         unreachable!()
@@ -1292,129 +856,5 @@ impl Session {
     pub fn get_total_token_usage(&self) -> Result<Option<i32>> {
         let metadata = self.get_metadata()?;
         Ok(metadata.total_tokens)
-    }
-
-    /// Display enhanced context usage with session totals
-    pub async fn display_context_usage(&self) -> Result<()> {
-        let provider = self.agent.provider().await?;
-        let model_config = provider.get_model_config();
-        let context_limit = model_config.context_limit.unwrap_or(32000);
-
-        match self.get_metadata() {
-            Ok(metadata) => {
-                let total_tokens = metadata.total_tokens.unwrap_or(0) as usize;
-
-                output::display_context_usage(total_tokens, context_limit);
-            }
-            Err(_) => {
-                output::display_context_usage(0, context_limit);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle prompt command execution
-    async fn handle_prompt_command(&mut self, opts: input::PromptCommandOptions) -> Result<()> {
-        // name is required
-        if opts.name.is_empty() {
-            output::render_error("Prompt name argument is required");
-            return Ok(());
-        }
-
-        if opts.info {
-            match self.get_prompt_info(&opts.name).await? {
-                Some(info) => output::render_prompt_info(&info),
-                None => output::render_error(&format!("Prompt '{}' not found", opts.name)),
-            }
-        } else {
-            // Convert the arguments HashMap to a Value
-            let arguments = serde_json::to_value(opts.arguments)
-                .map_err(|e| anyhow::anyhow!("Failed to serialize arguments: {}", e))?;
-
-            match self.get_prompt(&opts.name, arguments).await {
-                Ok(messages) => {
-                    let start_len = self.messages.len();
-                    let mut valid = true;
-                    for (i, prompt_message) in messages.into_iter().enumerate() {
-                        let msg = Message::from(prompt_message);
-                        // ensure we get a User - Assistant - User type pattern
-                        let expected_role = if i % 2 == 0 {
-                            mcp_core::Role::User
-                        } else {
-                            mcp_core::Role::Assistant
-                        };
-
-                        if msg.role != expected_role {
-                            output::render_error(&format!(
-                                "Expected {:?} message at position {}, but found {:?}",
-                                expected_role, i, msg.role
-                            ));
-                            valid = false;
-                            // get rid of everything we added to messages
-                            self.messages.truncate(start_len);
-                            break;
-                        }
-
-                        if msg.role == mcp_core::Role::User {
-                            output::render_message(&msg, self.debug);
-                        }
-                        self.messages.push(msg);
-                    }
-
-                    if valid {
-                        output::show_thinking();
-                        self.process_agent_response(true).await?;
-                        output::hide_thinking();
-                    }
-                }
-                Err(e) => output::render_error(&e.to_string()),
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Save a recipe to a file
-    ///
-    /// # Arguments
-    /// * `recipe` - The recipe to save
-    /// * `filepath_str` - The path to save the recipe to
-    ///
-    /// # Returns
-    /// * `Result<PathBuf, String>` - The path the recipe was saved to or an error message
-    fn save_recipe(
-        &self,
-        recipe: &goose::recipe::Recipe,
-        filepath_str: &str,
-    ) -> anyhow::Result<PathBuf> {
-        let path_buf = PathBuf::from(filepath_str);
-        let mut path = path_buf.clone();
-
-        // Update the final path if it's relative
-        if path_buf.is_relative() {
-            // If the path is relative, resolve it relative to the current working directory
-            let cwd = std::env::current_dir().context("Failed to get current directory")?;
-            path = cwd.join(&path_buf);
-        }
-
-        // Check if parent directory exists
-        if let Some(parent) = path.parent() {
-            if !parent.exists() {
-                return Err(anyhow::anyhow!(
-                    "Directory '{}' does not exist",
-                    parent.display()
-                ));
-            }
-        }
-
-        // Try creating the file
-        let file = std::fs::File::create(path.as_path())
-            .context(format!("Failed to create file '{}'", path.display()))?;
-
-        // Write YAML
-        serde_yaml::to_writer(file, recipe).context("Failed to save recipe")?;
-
-        Ok(path)
     }
 }
